Move CMatchState traits
