+++ include/football/CMatchState.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "IJsonable.h"
-#include "traits/json.h"
+#include "football/traits/CMatchState.h"
 
 #include "types/CMatchState.h"
 
@@ -12,25 +12,7 @@
 
 #include "ExceptionUtils.h"
 
-namespace futsim
-{
-
-namespace football
-{
-class CMatchState;
-}
-
-template <> struct json_traits<football::CMatchState>
-{
-	//! JSON key for the class.
-	static inline constexpr std::string_view KEY = "Match state";
-	//! JSON key for the \copybrief football::CMatchState::mMandatoryPlayTimeState
-	static inline constexpr std::string_view MANDATORY_PERIOD_STATES_KEY = "Mandatory time period states";
-	//! JSON key for the \copybrief football::CMatchState::mExtraTimeState
-	static inline constexpr std::string_view EXTRA_PERIOD_STATES_KEY = "Extra time period states";
-};
-
-namespace football
+namespace futsim::football
 {
 
 /**
@@ -134,6 +116,4 @@ CMatchState::CMatchState(
 }
 FUTSIM_CATCH_AND_RETHROW_EXCEPTION( std::invalid_argument, "Error creating the match state." )
 
-} // football namespace
-
-} // futsim namespace
+} // futsim::football namespace
